refactor: make primitives and change event card
