+++ package.json
@@ -24,6 +24,7 @@
         "prettier": "^2.4.1",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
+        "react-icons": "^4.3.1",
         "swr": "^1.0.1",
         "theme-ui": "^0.12.0"
     },
@@ -34,10 +35,10 @@
         "@storybook/addon-links": "^6.1.14",
         "@storybook/preset-scss": "^1.0.3",
         "@storybook/react": "^6.1.14",
+        "@types/luxon": "^2.0.9",
         "@types/node": "^14.14.13",
         "@types/react": "^17.0.0",
         "@types/react-dom": "^17.0.0",
-        "@types/luxon": "^2.0.9",
         "@typescript-eslint/eslint-plugin": "^4.9.1",
         "@typescript-eslint/parser": "^4.9.1",
         "chromatic": "^6.0.5",
