--- src/components/particles/Box.tsx
@@ -1,4 +1,0 @@
-import { styled } from "@theme";
-
-const Box = styled("div");
-export default Box;
