+++ src/components/primitives.ts
@@ -1,0 +1,4 @@
+import { styled } from "@theme";
+
+export const Box = styled("div");
+export const Flex = styled(Box, { display: "flex" });
